cyberbot-0.3.7 Updates
Makes PWR = 1 light very dim if PWR switch is set to 0 and micro:bit is connected to USB.

Prevents micro:bit from displaying errors if PWR is switched from 1 to 0 while in a loop and USB connected.

Restarts program whenever PWR is set to 0 and back to 1, not just when USB is disconnected.  Does not require user to press micro:bit reset button.

Saves enough space to increase dance routine from 35 to 95 maneuvers.  Other tests indicate that 23 frequency_out calls can be added to a (previously-full) TV remote program, and breadboard LEDs can now be added to previously full whiskers navigation program with room to spare for a second sensor subsystem.