Production Firmware, Organize for releases, and API updates where 2 pins are parameters
Production firmware: v1.0.2

Organize for releases
  Release
    cyberbot-firmware-update-v1-0-2
    cyberbot-micropython-v0-5-0
  Dev
    firmware
    loader
    makeCode
    Tests

API updates where 2 pins are parameters

In methods with a second pin, that pin was moved from the method parameters to the object properties.

Example:
  bot(irDetPin).ir_detect(irLedPin, frequency)
is now
  bot(irLedPin, irDetPin).ir_detect(frequency)

In methods where a group of pins are denoted, higher pin should come first.

Example
  io(largerPin).states(smallerPin,pinStates)
is now
  io(largerPin, smallerPin).states(pinStates)